requirements.txt
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.26.0
flask>=3.0.0
plotly>=5.18.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pytest>=7.4.0
